requirements.txt
mkdocs==1.6.1
mkdocs-material==9.6.14
mkdocs-mermaid2-plugin==1.2.1
mkdocs-drawio-exporter==0.10.2
mkdocs-confluence-publisher==0.1.2
mkdocs-minify-plugin==0.7.1
mkdocs-git-revision-date-localized-plugin==1.2.2
python-dotenv==1.1.0
pillow==10.2.0
cairosvg==2.7.1
